requirements.txt
# Requirements for FastVLM with Qwen3 LLM backbone
# Qwen3 requires transformers >= 4.51.0 but NOT 5.x

transformers>=4.51.0,<5.0.0
torch>=2.1.0
torchvision>=0.16.0
accelerate>=0.26.0
peft>=0.10.0
bitsandbytes>=0.43.0   # for QLoRA training
sentencepiece
einops
timm>=0.9.0            # for FastViT backbone

# Additional dependencies
Pillow>=9.0.0
numpy>=1.21.0
tokenizers>=0.14.0

# Optional: for training
deepspeed>=0.12.0

# Optional: for iOS/CoreML export
coremltools>=7.0
